// src/modules/bitrix/application/interfaces/activities/activities.interface.ts
export interface B24Activity {
  ID: string;
  OWNER_ID: string;
  OWNER_TYPE_ID: string;
  TYPE_ID: string;
  PROVIDER_ID: string;
  PROVIDER_TYPE_ID: string;
  PROVIDER_GROUP_ID: string | null;
  ASSOCIATED_ENTITY_ID: string;
  SUBJECT: string;
  CREATED: string;
  LAST_UPDATED: string;
  START_TIME: string;
  END_TIME: string;
  DEADLINE: string;
  COMPLETED: string;
  STATUS: string;
  RESPONSIBLE_ID: string;
  PRIORITY: string;
  NOTIFY_TYPE: string;
  NOTIFY_VALUE: string;
  DESCRIPTION: string;
  DESCRIPTION_TYPE: string;
  DIRECTION: string;
  LOCATION: string;
  SETTINGS: any[];
  ORIGINATOR_ID: string | null;
  ORIGIN_ID: string | null;
  AUTHOR_ID: string;
  EDITOR_ID: string;
  PROVIDER_PARAMS: any[];
  PROVIDER_DATA: string | null;
  RESULT_MARK: string;
  RESULT_VALUE: string | null;
  RESULT_SUM: string | null;
  RESULT_CURRENCY_ID: string | null;
  RESULT_STATUS: string;
  RESULT_STREAM: string;
  RESULT_SOURCE_ID: string | null;
  AUTOCOMPLETE_RULE: string;
}